requirements.txt
# ERIS - Economic Regime Intelligence System
# Python 3.10+

# Data collection & HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
newsapi-python>=0.2.7
yfinance>=0.2.28
fredapi>=0.5.0
praw>=7.7.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
trafilatura>=1.6.0
datasketch>=1.6.0
langdetect>=1.0.9
spacy>=3.7.0
schedule>=1.2.0

# NLP & ML
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
bertopic>=0.16.0
umap-learn>=0.5.0
scikit-learn>=1.3.0
xgboost>=2.0.0
hmmlearn>=0.3.0
ruptures>=1.1.9
shap>=0.43.0

# Statistics & validation
statsmodels>=0.14.0

# LLM
openai>=1.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Visualization
plotly>=5.18.0
altair>=5.0.0
matplotlib>=3.7.0

# Web app
streamlit>=1.28.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
ruff>=0.1.0
